refactor(network): optimized client mem usage
Use Read() instead of Peek()+Skip() in ReadByte() and appendBodyFixedSize()
to avoid holding references to the underlying buffer. Peek() returns slices
that reference the internal buffer, preventing it from being released or
reused until the reference is dropped.

Also updated cloudwego/gopkg dependency and moved connstate import to
gopkg/net/connstate.
