+++ .codecov.yml
@@ -13,3 +13,6 @@ coverage:
         # pass if coverage drops by no more than 0.05%
         # this is possibly caused by unstable coverage.
         threshold: 0.05%
+    patch:
+      default:
+        target: 80%
