+++ pkg/network/standard/transport.go
@@ -26,7 +26,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/cloudwego/gopkg/connstate"
+	"github.com/cloudwego/gopkg/net/connstate"
 
 	"github.com/cloudwego/hertz/pkg/common/config"
 	"github.com/cloudwego/hertz/pkg/common/hlog"
